+++ rpc/ethereum/namespaces/debug/api.go
@@ -19,8 +19,10 @@ import (
 	evmtypes "github.com/tharsis/ethermint/x/evm/types"
 
 	"github.com/cosmos/cosmos-sdk/client"
+	stderrors "github.com/pkg/errors"
 
 	"github.com/cosmos/cosmos-sdk/server"
+
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/consensus/ethash"
@@ -351,7 +353,10 @@ func (a *API) StartCPUProfile(file string) error {
 		}
 		if err := pprof.StartCPUProfile(f); err != nil {
 			a.logger.Debug("cpu profiling already in use", "error", err.Error())
-			f.Close()
+			if err := f.Close(); err != nil {
+				a.logger.Debug("failed to close cpu profile file")
+				return stderrors.Wrap(err, "failed to close cpu profile file")
+			}
 			return err
 		}
 
@@ -375,7 +380,10 @@ func (a *API) StopCPUProfile() error {
 	case a.handler.cpuFile != nil:
 		a.logger.Info("Done writing CPU profile", "profile", a.handler.cpuFilename)
 		pprof.StopCPUProfile()
-		a.handler.cpuFile.Close()
+		if err := a.handler.cpuFile.Close(); err != nil {
+			a.logger.Debug("failed to close cpu file")
+			return stderrors.Wrap(err, "failed to close cpu file")
+		}
 		a.handler.cpuFile = nil
 		a.handler.cpuFilename = ""
 		return nil